+++ lib/modules/forms/forms_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: forms
-// Updated at: 2026-05-27 21:59:41
-// Commit: feat(forms): add bi-directional event bus handler for forms
+// Updated at: 2026-05-28 22:25:49
+// Commit: feat(forms): integrate custom telemetry logger in forms
 
 class formsComponentState {
-  final String commitRef = '1532';
-  final String lastMessage = 'feat(forms): add bi-directional event bus handler for forms';
-  final DateTime timestamp = DateTime.parse('2026-05-27T21:59:41+05:00');
+  final String commitRef = '1561';
+  final String lastMessage = 'feat(forms): integrate custom telemetry logger in forms';
+  final DateTime timestamp = DateTime.parse('2026-05-28T22:25:49+05:00');
 }