+++ CHANGELOG.md
@@ -1562,3 +1562,4 @@
 - [2026-05-28 20:30] perf(perf): optimize memory consumption in perf list viewport
 - [2026-05-28 20:42] feat(analytics): integrate custom telemetry logger in analytics
 - [2026-05-28 21:46] test(theme): implement widget integration tests for theme
+- [2026-05-28 22:25] feat(forms): integrate custom telemetry logger in forms